Move RDFUtils into API
